+++ stroom-data/stroom-data-store-impl-fs/src/main/java/stroom/data/store/impl/fs/FsVolumeService.java
@@ -180,7 +180,7 @@ public class FsVolumeService implements EntityEvent.Handler, Clearable, Flushabl
     }
 
     private ResultPage<FsVolume> doFind(final FindFsVolumeCriteria criteria) {
-        return ResultPage.createCriterialBasedList(fsVolumeDao.find(criteria), criteria);
+        return fsVolumeDao.find(criteria);
     }
 
     FsVolume getVolume() {